PHP 8.0 | Arrays::getDoubleArrowPtr(): handle attributes
As of PHP 8.0, attributes can be attached to functions, closures, arrow functions, classes, interfaces, traits, class constants, properties, methods and method parameters.

Attributes effectively instantiate a new instance of the class being referenced and can pass arguments to the constructor of that class.

As, while uncommon, closures or arrow functions _could_ be used in an array key, there _could_ be an attribute attached to these.

This adjusts the method to skip over attributes to prevent them confusing the double arrow pointer determination.

Includes tests.

Refs:
* https://wiki.php.net/rfc/attributes_v2