Utils\TextStrings: remove support for PHPCS < 3.7.1
Includes removing the `getEndOfDoubleQuotedString()` method, which is no longer needed.

The method can be safely removed as it was never in a tagged release.

Note: the tests for the `TextStrings::getEndOfDoubleQuotedString()` method have been re-purposed as additional safeguards for the `TextStrings::getEndOfCompleteTextString()` method.

Also note that while the caching in the `TextStrings::getEndOfCompleteTextString()` method may be a bit over the top now, it will still be helpful when handling source documents which contain long multi-line text strings, so better to keep it in place.