BackCompat\BCTokens: remove the `tokenName()` method
Follow up on 250 and 347.

PR 250 introduced a polyfill for the PHPCS native `Tokens::tokenName()` method as this method was only available in PHPCS 3.0.0 and higher.

However, after the version drop, this method is no longer needed as the PHPCS native functionality can now be used in all cases.
This was missed in the version drop PR 347.

The method can be safely removed as it was never in a tagged release.

Note: as PHPCS itself does not have any tests for the native method, I'm keeping the associated test class and moving it to a separate `Xtra` directory for tests not directly related to PHPCSUtils. These tests should probably be pulled upstream at some point.