UtilityMethodTestCase::getTargetToken(): throw an exception when delimiter not found
PR 273 updated the  UtilityMethodTestCase::getTargetToken()``method to fail a test with a descriptive error messages when the test marker/delimiter comment could not be found.

This PR is a follow-up to that and instead of outright failing the test, the method will now throw a new, PHPCSUtils native `PHPCSUtils\Exceptions\TestMarkerNotFound` exception, which extends the PHP native `OutOfBoundsException`.

Note: the previous change has not been in a tagged release yet.

The reason for making this change is as follows:
1. The test run will still be marked as unsuccessful, so in that sense, there is no change.
2. Marking the test as "errored" is semantically more correct as a basic condition for the test to be able to run (precondition for finding the target token) has not been fulfilled.

Includes perfunctory tests for the new `TestMarkerNotFound` exception.
Includes updating the related test in the `GetTargetTokenTest` class to expect the new exception.