QA: make all classes `final`
The "class vs sniff name" matcher in the PHPCS autoloader does not handle sniffs classes extending other sniffs well.

Now, while this "standard" doesn't actually contain any sniffs, the utilities in this library are not intended to be extended anyway, aside from the explicitly `abstract` base sniffs.

So, with that in mind, I'm making all classes in this library `final`, with the exception of some select test classes (the `BCFile` ones as those need to be extended to safeguard that the PHPCSUtils native functionality mirrors the PHPCS native functionality).

If anyone has issues with any particular class now being `final`, please report this ASAP and please include a good use-case of why that particular class should not be `final`.