FunctionDeclarations::getParameters(): add caching
Follow up on 332

While most function declarations will be quite straight-forward, for function declarations with lots of parameters, this method can be token walking intensive.

As this method is also used by the `Operators::isReference()` method and in the future potentially by yet another method, caching the results of the function seems prudent.

Includes a dedicated test to verify the cache is used and working.