Arrays::isShortArray()/Lists::isShortList(): consolidate the caching [2]
This is step 2 in a refactor to improve short list/short array determination.

While caching had previously already been implemented, as things were, there could be four caches which would each refer to the same result, i.e.
1. A cache for an opener token passed to `Arrays::isShortArray()`.
2. A cache for an closer token of the same set of brackets passed to `Arrays::isShortArray()`.
3. A cache for an opener token of the same set of brackets passed to `Lists::isShortList()`.
4. A cache for an closer token of the same set of brackets passed to `Lists::isShortList()`.

This also meant that, depending on how the methods were called, the underlying `IsShortArrayOrList` class could be called up to four times to determine what a set of brackets signified, while the result would already be known after the first time and is not subject to change.

This commit introduces a wrapper class around the `IsShortArrayOrList` class which will handle the caching and will only create one cache entry for all four potential starting points, making the caching far more efficient and effective.

Includes adding a set of dedicated tests to cover the functionality in the `IsShortArrayOrListWithCache` class, replacing the previous caching tests in the `IsShortArrayOrListTest` class.