IsShortArrayOrList: only accept bracket opener tokens
As the new `IsShortArrayOrListWithCache` class will now be the entry point to the `IsShortArrayOrList` class, we can rely upon the token being passed to the `IsShortArrayOrList` always being an _opener_ for a complete set of brackets and can remove some redundancy from the `IsShortArrayOrList` class.

Includes making the exception for when an incorrect token is passed more specific.

Includes updating the `IsShortArrayOrList` tests to no longer pass closer tokens to the `IsShortArrayOrList` class.