IsShortArrayOrList::isInForeach(): improve logic
This commit improves the stability of the `IsShortArrayOrList::isInForeach()` method by:
* Using the `Context::inForeachCondition()` method to determine which part of a `foreach` condition we are looking at.
* Handling non-nested short arrays in the "before as" part of the condition, which should improve performance a little.
* Ensuring that only non-nested short lists in the "after as" part of the condition are handled by this method.
    Anything nested should be handled based on the _outer_ set of brackets as otherwise the results will be unreliable.

This also prevents incorrect results for code like this `foreach(["a"=>[]] as $v)`, where the empty array in the "before as" would previously be identified as a short list due to the double arrow before it.

Includes additional tests.

Refs:
* PHPCompatibility/PHPCompatibility 1341