UtilityMethodTestCase::getTargetToken(): change to a static method
The method effectively was already `static` as it didn't use `$this`. Making it explicitly static will allow for using it in "set up before class" test fixture methods.

This is particularly useful when a "cache" of information would need to be set up using the same code as the code under test.
If this is done in a "set up" fixture, the code being run would be counted for the code coverage calculation, while when don in "set up before class", the code being run will _not_ be counted for code coverage, allowing for cleaner coverage reports.

Includes introducing some minor variations in the pre-existing tests to safeguard that both static calls as well as non-static calls to the method are supported.