Collections: update array/list token collection for PHPCS 3.7.2
PHPCS 3.7.2 has been released with a fix related to the tokenization of short arrays/short lists.

With that fix included, sniffs no longer need to examine square bracket tokens anymore to check if those happen to be a short list/short array (well, not until the next bug is discovered).

This commit updates all BC compatible token collections related to short lists/short arrays to only include the `T_[OPEN|CLOSE]_SQUARE_BRACKET` tokens when on PHPCS < 3.7.2.

The net effect of this change should be that sniffing will be significantly faster on PHPCS 3.7.2.

Includes adjusted unit tests for the Collections.

**Note: _if/when the minimum supported PHPCS version goes up to PHPCS 3.7.2 or higher, this commit can be reverted and the square bracket tokens can be removed from the property based arrays instead._**

Ref: squizlabs/PHP_CodeSniffer 3632