Internal/IsShortArrayOrList[WithCache]: use stable token collections
The token collections from the `Collections` class can return different tokens depending on the PHPCS version used.
This can impact the functionality of various utilities in PHPCSUtils.

While for _most_ functionality, this is exactly the intended behaviour, for the code in the `Internal` classes, it is not and that code should still support the _full_ functionality (determining whether a square bracket is a real square bracket, a short array or a short list) as otherwise the code becomes untestable (and more difficult to maintain).

Sniffs should use the `Collections` class token arrays and should - depending on the PHPCS version - just no longer send "plain square brackets" to these classes, which will optimize sniff performance.