Lists::getAssignments(): bug fix for exotic keys
If function calls or arrays would be used as the _key_ for a list item and the call or array would contain a comma, the return value for the `Lists::getAssignments()` method would be incorrect and would identify more list items than there really are.

Fixed now.

Includes additional tests.