AbstractArrayDeclarationSniff: extra defensive coding for live coding/parse errors
In case of live coding/an unfinished array, the `PassedParameters::getParameters()` will return an empty array (no items), while in reality, it couldn't be determined.

To prevent incorrect sniff results, it is better for the sniff to bow out early in those situations.
This prevents a "Trying to access array offset on value of type bool" PHP notice.

Fixed now.

Includes additional test.