GH Actions: switch to Coveralls action runner to upload reports
Simplify the code coverage workflow by removing the dependency on the `php-coveralls/php-coveralls` package and switching to the `coverallsapp/github-action` action runner, which, as of the release of the [0.6.5 version of the Coverage Reporter](https://github.com/coverallsapp/coverage-reporter/releases/tag/v0.6.5) now natively supports the Clover format.

It is not yet clear whether this also fixes the issues with Dependabot and PRs from forks. If not, the `github-token: ${{ secrets.COVERALLS_TOKEN }}` should be removed in favour of allowing the GitHub native token to be used.