Tests/GetMethodPropertiesTest: sync with upstream [1] - use named items in data sets
These tests, for the most part, already had data set names in PHPCSUtils.

This commit adds the parameter name for each item in the data set, this time in an effort to make it more straight forward to update and add tests as it will be more obvious what each key in the data set signifies.

Includes fixing the data types in the docblocks and making them more specific, where relevant.