Composer: update the PHPUnit Polyfills to v 3.0
A PHPUnit 11 compatible version of the PHPUnit Polyfills has been released, so let's start using it.

Mind: PHPUnit 11 throws deprecation notices about the use of annotations instead of attributes.
On PHPUnit < 10.5.32 and PHPUnit < 11.3.3, these deprecation notices would fail the test run.

PHPUnit 10.5.32 and 11.3.3 introduce two new options for both the configuration file and as CLI arguments: `displayDetailsOnPhpunitDeprecations` and `failOnPhpunitDeprecation`, both of which will now default to `false`.

These options have been added to the PHPUnit configuration for PHPUnit 10/11 to safeguard them against changes in the default value.
This does mean that the configuration is no longer compatible with PHPUnit 10.0 - 10.5.31 and 11.0 - 11.3.2 and would fail the test run on seeing those config options.

With this in mind, it could be considered to set an explicit PHPUnit dependency to ensure the test suite isn't run on lower PHPUnit 10/11 versions, but as that is unlikely anyway, as Composer should always install the most recent version of PHPUnit, I deem this unnecessary.

Ref:
* https://github.com/Yoast/PHPUnit-Polyfills/releases/tag/3.0.0