Tests/PolyfilledTestCase: make compatible with PHPUnit Polyfills 3.x
As the available polyfill traits are different between the 1.x, 2.x and 3.x versions of the PHPUnit Polyfills, a different class definition is needed for PHPUnit Polyfills 3, which removes two traits (one used in this class) and introduces three new traits.

This commit adds the version-conditional class declaration for PHPUnit Polyfills 3.0.

Ref:
* https://github.com/Yoast/PHPUnit-Polyfills/releases/tag/3.0.0