Tests: data set keys should match parameter names
Initially to prevent issues with a few PHPUnit versions which were around in the early days of PHP 8.0, where these would accidentally be seen as parameter names.

So, as a best practice, when using keys in data sets, the keys should match the parameter names of the receiving test method.

And if a "super-"dataprovider is used and individual tests only need a subset of the keys from the data provider, add dedicated dataprovider for the individual test to remove the unneeded keys.

An ulterior reason to make this change is that PHPUnit 11 will always treat the keys as parameter names, so this can be seen as a preparation step for allowing for PHPUnit 11 support.