Tests: remove redundant `@group` annotations
The `@group` annotations were initially added to:
- allow for running closely related BCFile tests + Utils tests in one go;
- for running other closely related tests together;
- and for selectively excluding some unstable tests from CI test runs.

For everything else they should be regarded as redundant as the PHPUnit native `--filter` argument can sufficiently handle with running a group of tests from a single directory/namespace in modern PHPUnit versions (and who still runs locally on older ones ?).

This commit removes `@group` tags which are not needed based on that reasoning.