PHP 8.4 | BackCompat\Helper: fix implicitly nullable types deprecation notice
The `$phpcsFile` parameter is optional for the `getEncoding()` and the `ignoreAnnotations` method, but the type declaration cannot be made nullable as the minimum supported PHP version of this package is PHP 5.4, while PHP 7.1 is needed for the nullable operator.

Removing the `File` type declaration effectively widens the type to `mixed`, which fixes the deprecation notice.

As the class is `final`, this change does not constitute a BC-break.

Ref: https://wiki.php.net/rfc/deprecate-implicitly-nullable-types