IsReferenceTest: fix invalid test case
Sister-PR to PHPCSStandards/PHP_CodeSniffer 426

Closures use clauses only take plain variables, not complex variables, like properties or array keys.

In other words, this test case as-is, was a parse error.

For the purposes of this test, it makes no difference what type of variable is passed, so let's fix the test.