Tests: rename `AssertAttributeSame` trait to `AssertPropertySame`
The name was originally based on the name of the assertion available in PHPUnit, however, what with the introduction of the language construct named attributes in PHP 8.0, the name is now not very helpful/clear.

What this trait really does, is check the value of an object property, so let's make that more obvious in the naming.

Note: as this is an internal test-only file, this rename is not a breaking change.