UtilityMethodTestCase: new `parseFile()` method
This abstracts the parsing of a file out from the `setUpTestFile()` method, without changing the functionality.

Having the parsing of a file as a separate method allows for tests to parse a secondary test case file for use in a test.

As the new method doesn't create any functional changes, it is already covered by existing tests.