:sparkles: New `ContextTracking\NamespaceTracker`
New tracker class which keeps track of the current namespace.

Useful for sniffs which need an awareness of the current namespace during their checks, or only need to run in a namespaced file or visa versa.

See the class docblock for an implementation example.

Use the `NamespaceTracker::getNamespace()` method to retrieve the applicable namespace for any token in a file.

Alternatively, use the `NamespaceTracker::getNamespaceInfo()` method to retrieve the stack pointer to the effective start and end of the applicable namespace + the namespace name for any token in a file.

Note: this does not imply that the token passed is _subject_ to namespacing.

This tracker is highly optimized for both performance as well as accuracy and will, when used correctly, only walk a file once, no matter how often sniffs request the namespace or how many sniffs use the Tracker.

Includes extensive unit tests for the tracker.