WIP - needs clean up of Tracker, plus double-check + update commit msg | :sparkles: New `ContextTracking\ImportUseTracker`
New tracker class which keeps track of all seen import use statements.

Useful for sniffs which need an awareness of which import use statements are in effect.

See the class docblock for an implementation example.

Use the `getUseStatement()` method to retrieve the applicable import use statements for any token in a file.

Alternatively, use the `getUseStatementsInfo()` method to retrieve the stack pointer as of where the current set of statements is applicable + the import use statements for any token in a file.

This tracker is highly optimized for both performance as well as accuracy and will, when used correctly, only walk a file once, no matter how often sniffs request the use statements.

Includes extensive unit tests for the tracker.