:sparkles: New `ContextTracking\DeclaredFunctionsTracker`
New tracker class which keeps track of all named global/namespaced functions - NOT methods! - declared in a file.

Useful for sniffs which need to find a function declaration in a file, like sniffs working with callbacks.

See the class docblock for an implementation example.

Use the `DeclaredFunctionsTracker::findInFile(File $phpcsFile, $fqnFunctionName): int|false` method to retrieve the stack pointer to the `T_FUNCTION` keyword for the function declaration. Returns `false` if the function is not declared in the file.

Alternatively (discouraged), use the `DeclaredFunctionsTracker::getFunctions()` method to retrieve an array of all functions declared in the file.

This tracker is optimized, as much as possible, for performance, while not compromising accuracy and will, when used correctly, only walk a file once, no matter how often sniffs request the functions or how many sniffs use the Tracker.

Includes extensive unit tests for the tracker.