PHP 8.4 | Collections::functionCallTokens(): add `T_EXIT`
As of PHP 8.4, `exit` and `die` will become "special" function calls and when used as a constant, it will be transformed to a function call at compile time. See the RFC for full details.

With this in mind, it is appropriate to add `T_EXIT` to the `functionCallTokens()` and by inheritance to the `parameterPassingTokens()`.

Includes unit tests.

Ref: https://wiki.php.net/rfc/exit-as-function