:sparkles: New PHPCSUtils native `RuntimeException` class
This exception extends the PHPCS native `RuntimeException` to allow pre-existing PHPCSUtils code to switch to this exception without causing a breaking change.

This exception is also not `final` to allow more specific exceptions to extend from it. This is, again, deliberate, to allow pre-existing PHPCSUtils code to switch to more specific child-exceptions without causing a breaking change.

Includes changing the `InvalidTokenArray` exception to extend this new exception.