:sparkles: New `PHPCSUtils\Exceptions\OutOfBoundsStackPtr` exception class
New exception to flag a passed stack pointer which doesn't exist in the $phpcsFile.

This exception extends the PHPCSUtils native `RuntimeException` to allow pre-existing PHPCSUtils code to switch to this exception without causing a breaking change, though extending the PHPCSUtils `ValueError` class would also have been an option.

By rights, this exception should probably extend the PHP native `OutOfBoundsException` class, but not extending the `RuntimeException` would be a breaking change.

By the time a 2.0 release happens, it should be considered to switch the parent class for the exception to the PHP native `OutOfBoundsException`.

Includes perfunctory test for the exception.