:sparkles: New `PHPCSUtils\Exceptions\TypeError` exception class
New exception to flag an invalid argument type passed to a method using a standardized message.

This exception extends the PHPCSUtils native `RuntimeException` to allow pre-existing PHPCSUtils code to switch to this exception without causing a breaking change.

If, by the time a 2.0 release happens, usage of this exception has not been replaced with PHP native type declarations yet, it should be considered to switch the parent class for the exception to the PHP native `TypeError` class (PHP 7.0+), which this exception largely emulates.

Includes perfunctory test for the exception.