:sparkles: New `PHPCSUtils\Exceptions\ValueError` exception class
New exception to flag arguments using the correct type, but where the value doesn't comply with predefined restrictions, like an empty string being passed, when only a non-empty string is accepted or a negative integer being passed when a positive integer is expected.

The exception ensures these issues are flagged with a more consistent message format using a standardized message prefix.

This exception extends the PHPCSUtils native `RuntimeException` to allow pre-existing PHPCSUtils code to switch to this exception without causing a breaking change.

By rights, this exception should extend the PHP native `ValueError` class, but unfortunately, that class is not available until PHP 8.0 (and not extending the `RuntimeException` would be a breaking change).

By the time a 2.0 release happens, it should be considered to switch the parent class for the exception to the PHP native `ValueError` (PHP 8.0+) if the minimum supported PHP version allows for it.

Includes perfunctory test for the exception.