:sparkles: New `PHPCSUtils\Exceptions\UnexpectedTokenType` exception class
New exception to flag a passed stack pointer argument, which does not comply with the token type requirements of the receiving method.

This exception extends the PHPCSUtils native `RuntimeException` to allow pre-existing PHPCSUtils code to switch to this exception without causing a breaking change, though extending the PHPCSUtils `ValueError` class would also have been an option.

By rights, this exception should probably extend the PHP native `InvalidArgumentException` class, but not extending the `RuntimeException` would be a breaking change.

By the time a 2.0 release happens, it should be considered to switch the parent class for the exception to the PHP native `InvalidArgumentException`.

Includes perfunctory test for the exception.